format sections to utils
