+++ server/data/disruption/writeup/bus-replacements-disruption-writeup-author.ts
@@ -3,9 +3,7 @@ import { BusReplacementsDisruptionData } from "@/server/data/disruption/data/bus
 import { Disruption } from "@/server/data/disruption/disruption";
 import { DisruptionWriteup } from "@/server/data/disruption/writeup/disruption-writeup";
 import { DisruptionWriteupAuthor } from "@/server/data/disruption/writeup/disruption-writeup-author";
-import { LineSection } from "@/server/data/line-section";
-import { LineShapeNode } from "@/server/data/static/line-routes/line-shape";
-import { listifyAnd, unique } from "@dan-schel/js-utils";
+import { formatSections } from "@/server/data/disruption/writeup/utils";
 
 /** DisruptionWriteupAuthor for BusReplacementsDisruptionData. */
 export class BusReplacementsDisruptionWriteupAuthor extends DisruptionWriteupAuthor {
@@ -30,61 +28,4 @@ export class BusReplacementsDisruptionWriteupAuthor extends DisruptionWriteupAut
       "medium",
     );
   }
-}
-
-function formatSections(app: App, sections: LineSection[]): string {
-  // TODO: Make this outside this file, and write tests for it.
-  if (sections.length === 0) {
-    throw new Error("Must have at least one section.");
-  }
-
-  if (sections.length === 1 || allTheSame(sections)) {
-    const a = formatLineShapeNode(app, sections[0].a);
-    const b = formatLineShapeNode(app, sections[0].b);
-    return `from ${a} to ${b}`;
-  }
-
-  const commonNode = getCommonNode(sections);
-  if (commonNode != null) {
-    const common = formatLineShapeNode(app, commonNode.common);
-    const connections = commonNode.connections.map((n) =>
-      formatLineShapeNode(app, n),
-    );
-    return `from ${common} to ${listifyAnd(connections)}`;
-  }
-
-  const individualSections = sections.map((s) => {
-    const a = formatLineShapeNode(app, s.a);
-    const b = formatLineShapeNode(app, s.b);
-    return `${a} to ${b}`;
-  });
-  return `from ${listifyAnd(individualSections)}`;
-}
-
-function formatLineShapeNode(app: App, node: LineShapeNode): string {
-  if (node === "the-city") return "the city";
-  return app.stations.require(node).name;
-}
-
-function allTheSame(sections: LineSection[]): boolean {
-  const first = sections[0];
-  return sections.every(
-    (s) =>
-      (s.a === first.a && s.b === first.b) ||
-      (s.a === first.b && s.b === first.a),
-  );
-}
-
-function getCommonNode(
-  sections: LineSection[],
-): { common: LineShapeNode; connections: LineShapeNode[] } | null {
-  const allNodes = unique(sections.map((s) => [s.a, s.b]).flat());
-
-  const common = allNodes.find((n) =>
-    sections.every((s) => s.a === n || s.b === n),
-  );
-  if (common == null) return null;
-
-  const connections = allNodes.filter((n) => n !== common);
-  return { common, connections };
 }
